pyproject.toml
[tool.poetry]
name = "imogen"
version = "0.1.0"
description = ""
authors = ["joypauls <joypaulsen3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.1"
numpy = "^1.25.2"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
